+++ app/(connected)/(with-sidebar)/(with-chat)/mes-jeunes/creation-jeune/pole-emploi/CreationJeunePoleEmploiPage.tsx
@@ -1,32 +1,30 @@
+'use client'
+
 import { withTransaction } from '@elastic/apm-rum-react'
-import { GetServerSideProps } from 'next'
-import { useRouter } from 'next/router'
-import React, { useState } from 'react'
+import { useRouter } from 'next/navigation'
+import React, { ReactElement, useState } from 'react'
 
 import FormulaireJeunePoleEmploi from 'components/jeune/FormulaireJeunePoleEmploi'
-import { StructureConseiller } from 'interfaces/conseiller'
 import { JeunePoleEmploiFormData } from 'interfaces/json/jeune'
-import { PageProps } from 'interfaces/pageProps'
 import { AlerteParam } from 'referentiel/alerteParam'
 import { useAlerte } from 'utils/alerteContext'
 import useMatomo from 'utils/analytics/useMatomo'
 import { usePortefeuille } from 'utils/portefeuilleContext'
 
-interface PoleEmploiCreationJeuneProps extends PageProps {}
-
-function PoleEmploiCreationJeune(): JSX.Element {
+function PoleEmploiCreationJeune(): ReactElement {
   const router = useRouter()
   const [_, setAlerte] = useAlerte()
   const [portefeuille, setPortefeuille] = usePortefeuille()
 
-  const [creationError, setCreationError] = useState<string>('')
+  const [creationError, setCreationError] = useState<string>()
   const [creationEnCours, setCreationEnCours] = useState<boolean>(false)
 
   async function creerJeunePoleEmploi(
     newJeune: JeunePoleEmploiFormData
   ): Promise<void> {
-    setCreationError('')
+    setCreationError(undefined)
     setCreationEnCours(true)
+
     try {
       const { createCompteJeunePoleEmploi } = await import(
         'services/jeunes.service'
@@ -39,7 +37,7 @@ function PoleEmploiCreationJeune(): JSX.Element {
 
       setPortefeuille(portefeuille.concat(beneficiaireCree))
       setAlerte(AlerteParam.creationBeneficiaire, beneficiaireCree.id)
-      await router.push('/mes-jeunes')
+      router.push('/mes-jeunes')
     } catch (error) {
       setCreationError(
         (error as Error).message || "Une erreur inconnue s'est produite"
@@ -58,41 +56,6 @@ function PoleEmploiCreationJeune(): JSX.Element {
       creationEnCours={creationEnCours}
     />
   )
-}
-
-export const getServerSideProps: GetServerSideProps<
-  PoleEmploiCreationJeuneProps
-> = async (context) => {
-  const { default: withMandatorySessionOrRedirect } = await import(
-    'utils/auth/withMandatorySessionOrRedirect'
-  )
-  const sessionOrRedirect = await withMandatorySessionOrRedirect(context)
-
-  if (!sessionOrRedirect.validSession) {
-    return { redirect: sessionOrRedirect.redirect }
-  }
-
-  if (
-    sessionOrRedirect.session.user.structure !==
-      StructureConseiller.POLE_EMPLOI &&
-    sessionOrRedirect.session.user.structure !==
-      StructureConseiller.POLE_EMPLOI_BRSA
-  ) {
-    return {
-      redirect: {
-        destination: '/mes-jeunes',
-        permanent: false,
-      },
-    }
-  }
-
-  return {
-    props: {
-      pageTitle: `Portefeuille - Créer bénéficiaire`,
-      pageHeader: `Créer un compte bénéficiaire`,
-      returnTo: '/mes-jeunes',
-    },
-  }
 }
 
 export default withTransaction(
