+++ app/(connected)/(with-sidebar)/(without-chat)/mes-jeunes/[idJeune]/actions/[idAction]/modification/page.tsx
@@ -8,7 +8,7 @@ import {
   PageRetourPortal,
 } from 'components/PageNavigationPortals'
 import { StatutAction } from 'interfaces/action'
-import { StructureConseiller } from 'interfaces/conseiller'
+import { estUserPoleEmploi } from 'interfaces/conseiller'
 import {
   getAction,
   getSituationsNonProfessionnelles,
@@ -39,7 +39,7 @@ export default async function ModificationAction({
   params: ModificationParams
 }) {
   const { user, accessToken } = await getMandatorySessionServerSide()
-  if (user.structure === StructureConseiller.POLE_EMPLOI) notFound()
+  if (estUserPoleEmploi(user)) notFound()
 
   const [actionContent, situationsNonProfessionnelles, actionsPredefinies] =
     await Promise.all([
