+++ components/jeune/AjouterJeuneButton.tsx
@@ -9,10 +9,10 @@ interface AjouterJeuneButtonProps {
 export function getAjouterJeuneHref(structure: StructureConseiller): string {
   switch (structure) {
     case StructureConseiller.MILO:
-      return '/mes-jeunes/milo/creation-jeune'
+      return '/mes-jeunes/creation-jeune/milo'
     case StructureConseiller.POLE_EMPLOI:
     case StructureConseiller.POLE_EMPLOI_BRSA:
-      return '/mes-jeunes/pole-emploi/creation-jeune'
+      return '/mes-jeunes/creation-jeune/pole-emploi'
     default:
       return ''
   }
