+++ tests/api/fichiers.api.test.tsx
@@ -19,7 +19,7 @@ describe('GET api/fichier/[idFichier]', () => {
     })
 
     // When
-    const promise = GET(new Request('https://www.mozilla.org/favicon.ico'), {
+    const promise = GET(new Request('https://www.perdu.com'), {
       params: { idFichier: 'idFichier' },
     })
