+++ utils/route-labels.ts
@@ -15,7 +15,7 @@ const mapRoutesToLabels: Map<RegExp, string> = new Map<RegExp, string>([
 
   // Portefeuille
   [/\/mes-jeunes$/, 'Portefeuille'],
-  [/\/mes-jeunes\/(milo|pole-emploi)\/creation-jeune$/, 'Création'],
+  [/\/mes-jeunes\/creation-jeune\/(milo|pole-emploi)$/, 'Création'],
   [/\/mes-jeunes\/[\w-]+$/, labelFicheBeneficiaire],
   [/\/mes-jeunes\/[\w-]+\/favoris$/, labelFavoris],
   [/\/mes-jeunes\/[\w-]+\/informations$/, labelInformations],
